07-operasyonel-yazilimlar/kutuphane-yonetim-sistemi/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3
python-decouple==3.8
Pillow==10.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
drf-spectacular==0.26.5
djangorestframework-simplejwt==5.3.0
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
python-barcode==0.15.1
qrcode==7.4.2
reportlab==4.0.7
openpyxl==3.1.2